style(reports/landing): add larger scroll link area
