+++ src/pages/Reports/pages/Landing/components/TopSection/index.js
@@ -8,6 +8,7 @@ import BycicleParkingBgImg from '~/images/reports/bycicle-parking@3x.png';
 import BycicleParkingBgImgLargeScreen from '~/images/reports/landing-christin-hume-595752-unsplash.jpg';
 import Button from '~/components/Button';
 import Link from '~/components/Link';
+import ScrollLink from './ScrollLink';
 import { media } from '~/styles/utils';
 
 const Section = styled.section`
@@ -94,6 +95,7 @@ const TopSection = ({ toUrl }) => (
         <strong>Sagt uns wo</strong><br /> in 30 Sekunden
       </StyledLink>
     </StyledButton>
+    <ScrollLink />
   </Section>
 );
